Diverse Sonar fixes (1) Fjernet ubrukt parameter, maxThreads, i DokmotConsumer (2) Endret pakkenavn for altinnInntekt til altinninntekt for å være kompatibel med standard regex (3) Fjernet unødvendige deklareringer av RuntimeExceptions som RestClientException og ValidationException
