+++ apps/testnorge-inntekt/src/main/java/no/nav/registre/inntekt/consumer/rs/altinninntekt/dto/enums/AltinnEnum.java
@@ -1,4 +1,4 @@
-package no.nav.registre.inntekt.consumer.rs.altinnInntekt.dto.enums;
+package no.nav.registre.inntekt.consumer.rs.altinninntekt.dto.enums;
 
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import no.nav.registre.inntekt.utils.JsonAltinnEnumSerializer;
