+++ apps/testnorge-inntekt/src/main/java/no/nav/registre/inntekt/consumer/rs/altinninntekt/dto/enums/BegrunnelseIngenEllerRedusertUtbetalingKodeListe.java
@@ -1,4 +1,4 @@
-package no.nav.registre.inntekt.consumer.rs.altinnInntekt.dto.enums;
+package no.nav.registre.inntekt.consumer.rs.altinninntekt.dto.enums;
 
 public enum BegrunnelseIngenEllerRedusertUtbetalingKodeListe implements AltinnEnum {
     LOVLIG_FRAVAER("LovligFravaer"),
