+++ apps/testnorge-inntekt/src/main/java/no/nav/registre/inntekt/consumer/rs/altinninntekt/dto/rs/RsEndringIRefusjon.java
@@ -1,4 +1,4 @@
-package no.nav.registre.inntekt.consumer.rs.altinnInntekt.dto.rs;
+package no.nav.registre.inntekt.consumer.rs.altinninntekt.dto.rs;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.annotations.ApiModel;
