+++ apps/testnorge-inntekt/src/main/java/no/nav/registre/inntekt/consumer/rs/altinninntekt/dto/rs/RsPeriode.java
@@ -1,0 +1,27 @@
+package no.nav.registre.inntekt.consumer.rs.altinninntekt.dto.rs;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import io.swagger.annotations.ApiModel;
+import io.swagger.annotations.ApiModelProperty;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.NoArgsConstructor;
+import lombok.Value;
+
+import java.time.LocalDate;
+
+@ApiModel
+@Builder
+@Value
+@NoArgsConstructor(force = true)
+@AllArgsConstructor
+public class RsPeriode {
+
+    @JsonProperty
+    @ApiModelProperty(value = "Dato fra og med", example = "YYYY-MM-DD")
+    private LocalDate fom;
+    @JsonProperty
+    @ApiModelProperty(value = "Dato til og med", example = "YYYY-MM-DD")
+    private LocalDate tom;
+
+}
