+++ apps/testnorge-inntekt/src/main/java/no/nav/registre/inntekt/consumer/rs/altinninntekt/dto/rs/RsUtsettelseAvForeldrepenger.java
@@ -1,4 +1,4 @@
-package no.nav.registre.inntekt.consumer.rs.altinnInntekt.dto.rs;
+package no.nav.registre.inntekt.consumer.rs.altinninntekt.dto.rs;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.annotations.ApiModel;
@@ -7,19 +7,20 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.NoArgsConstructor;
 import lombok.Value;
+import no.nav.registre.inntekt.consumer.rs.altinninntekt.dto.enums.AarsakUtsettelseKodeListe;
 
 @ApiModel
 @Builder
 @Value
 @NoArgsConstructor(force = true)
 @AllArgsConstructor
-public class RsGraderingIForeldrepenger {
+public class RsUtsettelseAvForeldrepenger {
 
     @JsonProperty
-    @ApiModelProperty("Periode med foreldrepenger")
+    @ApiModelProperty("Periode for utsettelse av foreldrepenger")
     private RsPeriode periode;
     @JsonProperty
-    @ApiModelProperty("Heltall med arbeidstidsprosent")
-    private Integer arbeidstidprosent;
+    @ApiModelProperty()
+    private AarsakUtsettelseKodeListe aarsakTilUtsettelse;
 
 }
