+++ apps/testnorge-inntekt/src/main/java/no/nav/registre/inntekt/provider/rs/requests/AltinnInntektsmeldingRequest.java
@@ -6,7 +6,7 @@ import lombok.Value;
 
 import java.util.List;
 
-import no.nav.registre.inntekt.consumer.rs.altinnInntekt.dto.RsInntektsmeldingRequest;
+import no.nav.registre.inntekt.consumer.rs.altinninntekt.dto.RsInntektsmeldingRequest;
 import no.nav.registre.inntekt.consumer.rs.dokmot.dto.RsJoarkMetadata;
 
 @Value
