+++ apps/testnorge-inntekt/src/main/java/no/nav/registre/inntekt/provider/rs/v1/AltinnInntektControllerV1.java
@@ -43,7 +43,7 @@ public class AltinnInntektControllerV1 {
         return ResponseEntity.ok(response);
     }
 
-    private void validerInntektsmelding(AltinnInntektsmeldingRequest dollyRequest) throws ValidationException {
+    private void validerInntektsmelding(AltinnInntektsmeldingRequest dollyRequest) {
         for (var inntekt : dollyRequest.getInntekter()) {
             var arbeidsgiver = inntekt.getArbeidsgiver();
             var arbeidsgiverPrivat = inntekt.getArbeidsgiverPrivat();
