+++ apps/testnorge-inntekt/src/main/java/no/nav/registre/inntekt/provider/rs/v2/AltinnInntektControllerV2.java
@@ -42,7 +42,7 @@ public class AltinnInntektControllerV2 {
         return ResponseEntity.ok(response);
     }
 
-    private void validerInntektsmelding(AltinnInntektsmeldingRequest dollyRequest) throws ValidationException {
+    private void validerInntektsmelding(AltinnInntektsmeldingRequest dollyRequest) {
         for (var inntekt : dollyRequest.getInntekter()) {
             var arbeidsgiver = inntekt.getArbeidsgiver();
             var arbeidsgiverPrivat = inntekt.getArbeidsgiverPrivat();
