+++ apps/testnorge-inntekt/src/main/resources/application.properties
@@ -1,3 +1,2 @@
-dokmot.consumer.maxThreads=5
 application.name=testnorge-inntekt
 application.version=@application.version@
